# entity/user_api.go
package entity

type UserLoginReq struct {
	Username string `json:"username" binding:"required"`
	Password string `json:"password" binding:"required"`
}

type UserLoginRps struct {
	Token string `json:"token"`
}